refactor: decouple PR reviewer service from Azure DevOps using CodeReviewProvider
