+++ src/main/features/pr-reviewer/prReviewerService.ts
@@ -4,12 +4,6 @@ import os from 'os';
 import picomatch from 'picomatch';
 import matter from 'gray-matter';
 import { powerSaveBlocker } from 'electron';
-import * as azdev from 'azure-devops-node-api';
-import { IGitApi } from 'azure-devops-node-api/GitApi';
-import {
-  GitPullRequestSearchCriteria,
-  GitPullRequestCommentThread,
-} from 'azure-devops-node-api/interfaces/GitInterfaces';
 import { GitService } from '../../infrastructure/git/gitService';
 import { CopilotService } from '../../infrastructure/copilot/copilotService';
 import {
@@ -20,6 +14,7 @@ import {
 } from '../../../types';
 import { IssueTrackerProvider } from '../../infrastructure/providers/IssueTrackerProvider';
 import { DocumentationProvider } from '../../infrastructure/providers/DocumentationProvider';
+import { CodeReviewProvider } from '../../infrastructure/providers/CodeReviewProvider';
 import { createRequestDocumentationTool } from '../../infrastructure/copilot/tools/documentationTool';
 
 export function buildPhaseReviewPrompt(
@@ -182,6 +177,7 @@ export class PRReviewerService {
     private copilotService: CopilotService,
     private getIssueTrackerProvider: () => Promise<IssueTrackerProvider | null>,
     private getDocProvider: () => Promise<DocumentationProvider | null>,
+    private getCodeReviewProvider: () => Promise<CodeReviewProvider | null>,
   ) {}
 
   private extractUrls(text: string): string[] {
@@ -346,275 +342,28 @@ export class PRReviewerService {
     }
   }
 
-  parsePRUrl(url: string): {
-    org: string;
-    project: string;
-    repoName: string;
-    prNumber: number;
-  } | null {
-    const trimmed = url.trim();
-
-    // Pattern 1: dev.azure.com
-    // https://dev.azure.com/org/project/_git/repo/pullrequest/123
-    const devAzureRegex =
-      /https:\/\/dev\.azure\.com\/([^/]+)\/([^/]+)\/_git\/([^/]+)\/pullrequest\/(\d+)/i;
-    let match = devAzureRegex.exec(trimmed);
-    if (match) {
-      return {
-        org: match[1],
-        project: match[2],
-        repoName: match[3],
-        prNumber: parseInt(match[4]),
-      };
-    }
-
-    // Pattern 2: visualstudio.com
-    // https://org.visualstudio.com/project/_git/repo/pullrequest/123
-    const vsRegex =
-      /https:\/\/([^/]+)\.visualstudio\.com\/([^/]+)\/_git\/([^/]+)\/pullrequest\/(\d+)/i;
-    match = vsRegex.exec(trimmed);
-    if (match) {
-      return {
-        org: match[1],
-        project: match[2],
-        repoName: match[3],
-        prNumber: parseInt(match[4]),
-      };
-    }
-
-    return null;
-  }
-
-  parseRemoteUrl(url: string): {
-    org: string;
-    project: string;
-    repoName: string;
-  } | null {
-    const trimmed = url.trim();
-
-    // Pattern 1: HTTPS dev.azure.com
-    // https://dev.azure.com/org/project/_git/repo
-    // or https://user@dev.azure.com/org/project/_git/repo
-    const httpsRegex =
-      /https:\/\/(?:[^/]+@)?dev\.azure\.com\/([^/]+)\/([^/]+)\/_git\/([^/.]+)/i;
-    let match = httpsRegex.exec(trimmed);
-    if (match) {
-      return {
-        org: match[1],
-        project: match[2],
-        repoName: match[3],
-      };
-    }
-
-    // Pattern 2: SSH dev.azure.com
-    // git@ssh.dev.azure.com:v3/org/project/repo
-    const sshRegex = /git@ssh\.dev\.azure\.com:v3\/([^/]+)\/([^/]+)\/([^/.]+)/i;
-    match = sshRegex.exec(trimmed);
-    if (match) {
-      return {
-        org: match[1],
-        project: match[2],
-        repoName: match[3],
-      };
-    }
-
-    // Pattern 3: Legacy HTTPS visualstudio.com
-    // https://org.visualstudio.com/project/_git/repo
-    const vsRegex =
-      /https:\/\/([^/]+)\.visualstudio\.com\/([^/]+)\/_git\/([^/.]+)/i;
-    match = vsRegex.exec(trimmed);
-    if (match) {
-      return {
-        org: match[1],
-        project: match[2],
-        repoName: match[3],
-      };
-    }
-
-    return null;
-  }
-
-  private getOrgUrl(org: string): string {
-    if (org.startsWith('http://') || org.startsWith('https://')) {
-      return org;
-    }
-    return `https://dev.azure.com/${org}`;
-  }
-
   async getPRDetails(
     repoPath: string,
     prUrlOrId: string,
-    settings: AppSettings,
+    _settings: AppSettings,
   ): Promise<PRMetadata> {
-    let prNumber = parseInt(prUrlOrId);
-    const azureConn = settings.connectors?.azureDevOps;
-    let org = azureConn?.org || '';
-    let project = azureConn?.project || '';
-
-    const parsedUrl = this.parsePRUrl(prUrlOrId);
-    if (parsedUrl) {
-      prNumber = parsedUrl.prNumber;
-      org = parsedUrl.org;
-      project = parsedUrl.project;
-    } else if (isNaN(prNumber)) {
-      throw new Error(`Invalid Pull Request URL or ID format: "${prUrlOrId}"`);
-    } else {
-      // Try to detect org and project from remote URL
-      const remoteUrl = await this.gitService.getRemoteUrl(repoPath);
-      if (remoteUrl) {
-        const parsedRemote = this.parseRemoteUrl(remoteUrl);
-        if (parsedRemote) {
-          org = org || parsedRemote.org;
-          project = project || parsedRemote.project;
-        }
-      }
+    const provider = await this.getCodeReviewProvider();
+    if (!provider) {
+      throw new Error('No code review provider configured.');
     }
-
-    if (!org) {
-      throw new Error(
-        'Azure DevOps Organization is not configured in settings and could not be detected from git remote.',
-      );
-    }
-    if (!project) {
-      throw new Error(
-        'Azure DevOps Project is not configured in settings and could not be detected from git remote.',
-      );
-    }
-
-    const pat = settings.connectors?.azureDevOps?.pat;
-    if (!pat) {
-      throw new Error(
-        'Azure DevOps PAT token is missing. Please configure it in Settings.',
-      );
-    }
-
-    const orgUrl = this.getOrgUrl(org);
-    const authHandler = azdev.getPersonalAccessTokenHandler(pat);
-    const connection = new azdev.WebApi(orgUrl, authHandler);
-    const gitApi: IGitApi = await connection.getGitApi();
-
-    try {
-      const pr = await gitApi.getPullRequestById(prNumber);
-      if (!pr) {
-        throw new Error(`Pull Request #${prNumber} not found.`);
-      }
-
-      if (!pr.targetRefName) {
-        throw new Error(
-          `Pull Request #${prNumber} is missing target branch ref.`,
-        );
-      }
-
-      const cleanRef = (ref: string) => ref.replace(/^refs\/heads\//, '');
-      const orgUrl = this.getOrgUrl(org);
-      const baseUrl = orgUrl.endsWith('/') ? orgUrl.slice(0, -1) : orgUrl;
-      const prId = pr.pullRequestId?.toString() || prNumber.toString();
-      const repoName = pr.repository?.name || '';
-      const webUrl = `${baseUrl}/${project}/_git/${repoName}/pullrequest/${prId}`;
-
-      return {
-        id: prId,
-        title: pr.title || '',
-        description: pr.description || '',
-        sourceBranch: cleanRef(pr.sourceRefName || ''),
-        targetBranch: cleanRef(pr.targetRefName || ''),
-        author: pr.createdBy?.displayName || '',
-        repositoryName: repoName,
-        repositoryId: pr.repository?.id,
-        hostType: 'azure',
-        url: webUrl,
-      };
-    } catch (error: unknown) {
-      const errMsg = error instanceof Error ? error.message : String(error);
-      throw new Error(`Failed to fetch PR from Azure DevOps API: ${errMsg}`, {
-        cause: error,
-      });
-    }
+    const remoteUrl = await this.gitService.getRemoteUrl(repoPath);
+    return provider.getPRDetails(repoPath, prUrlOrId, remoteUrl);
   }
 
   async getProjectPRs(
     searchType: 'assigned' | 'created' | 'all',
-    settings: AppSettings,
+    _settings: AppSettings,
   ): Promise<PRMetadata[]> {
-    const azureConn = settings.connectors?.azureDevOps;
-    const org = azureConn?.org || '';
-    const project = azureConn?.project || '';
-    const pat = azureConn?.pat;
-
-    if (!org) {
-      throw new Error(
-        'Azure DevOps Organization is not configured in settings.',
-      );
+    const provider = await this.getCodeReviewProvider();
+    if (!provider) {
+      throw new Error('No code review provider configured.');
     }
-    if (!project) {
-      throw new Error('Azure DevOps Project is not configured in settings.');
-    }
-    if (!pat) {
-      throw new Error(
-        'Azure DevOps PAT token is missing. Please configure it in Settings.',
-      );
-    }
-
-    const orgUrl = this.getOrgUrl(org);
-    const authHandler = azdev.getPersonalAccessTokenHandler(pat);
-    const connection = new azdev.WebApi(orgUrl, authHandler);
-    const gitApi: IGitApi = await connection.getGitApi();
-
-    try {
-      let reviewerId: string | undefined;
-      let creatorId: string | undefined;
-
-      if (searchType === 'assigned' || searchType === 'created') {
-        const connectionData = await connection.connect();
-        const currentUserId = connectionData.authorizedUser?.id;
-        if (!currentUserId) {
-          throw new Error(
-            'Could not resolve current authenticated user identity ID.',
-          );
-        }
-        if (searchType === 'assigned') {
-          reviewerId = currentUserId;
-        } else {
-          creatorId = currentUserId;
-        }
-      }
-
-      const searchCriteria: GitPullRequestSearchCriteria = {
-        status: 1, // Active
-      };
-      if (reviewerId) searchCriteria.reviewerId = reviewerId;
-      if (creatorId) searchCriteria.creatorId = creatorId;
-
-      const prs = await gitApi.getPullRequestsByProject(
-        project,
-        searchCriteria,
-      );
-      const cleanRef = (ref: string) => ref.replace(/^refs\/heads\//, '');
-      return prs.map((pr) => {
-        const prId = pr.pullRequestId?.toString() || '';
-        const repoName = pr.repository?.name || '';
-        const baseUrl = orgUrl.endsWith('/') ? orgUrl.slice(0, -1) : orgUrl;
-        const webUrl = prId
-          ? `${baseUrl}/${project}/_git/${repoName}/pullrequest/${prId}`
-          : undefined;
-        return {
-          id: prId,
-          title: pr.title || '',
-          description: pr.description || '',
-          sourceBranch: cleanRef(pr.sourceRefName || ''),
-          targetBranch: cleanRef(pr.targetRefName || ''),
-          author: pr.createdBy?.displayName || '',
-          repositoryName: repoName,
-          hostType: 'azure',
-          url: webUrl,
-        };
-      });
-    } catch (error: unknown) {
-      const errMsg = error instanceof Error ? error.message : String(error);
-      throw new Error(`Failed to query PRs from Azure DevOps: ${errMsg}`, {
-        cause: error,
-      });
-    }
+    return provider.getProjectPRs(searchType);
   }
 
   getEffectiveRepoPath(repoPath: string): string {
@@ -769,7 +518,10 @@ export class PRReviewerService {
     if (expectedRepoName) {
       const remoteUrl = await this.gitService.getRemoteUrl(repoPath);
       if (remoteUrl) {
-        const parsedRemote = this.parseRemoteUrl(remoteUrl);
+        const provider = await this.getCodeReviewProvider();
+        const parsedRemote = provider
+          ? provider.parseRemoteUrl(remoteUrl)
+          : null;
         if (
           parsedRemote &&
           parsedRemote.repoName.toLowerCase() !== expectedRepoName.toLowerCase()
@@ -912,32 +664,13 @@ export class PRReviewerService {
         (phase) => phase.attach && phase.attach.toLowerCase().includes('story'),
       );
 
-      if (
-        anyPhaseNeedsStory &&
-        prDetails &&
-        prDetails.repositoryId &&
-        options.prId
-      ) {
+      if (anyPhaseNeedsStory && prDetails && options.prId) {
         try {
-          const prNumber = parseInt(prDetails.id);
-          const azureConn = settings.connectors?.azureDevOps;
-          let org = azureConn?.org || '';
-          let project = azureConn?.project || '';
-          const parsedUrl = this.parsePRUrl(options.prId);
-          if (parsedUrl) {
-            org = parsedUrl.org;
-            project = parsedUrl.project;
-          }
-          const pat = azureConn?.pat;
-          if (pat && org && project) {
-            const orgUrl = this.getOrgUrl(org);
-            const authHandler = azdev.getPersonalAccessTokenHandler(pat);
-            const connection = new azdev.WebApi(orgUrl, authHandler);
-            const gitApi: IGitApi = await connection.getGitApi();
-
-            const workItemRefs = await gitApi.getPullRequestWorkItemRefs(
+          const provider = await this.getCodeReviewProvider();
+          if (provider) {
+            const workItemRefs = await provider.getLinkedTickets(
+              prDetails.id,
               prDetails.repositoryId,
-              prNumber,
             );
 
             if (workItemRefs && workItemRefs.length > 0) {
@@ -945,16 +678,16 @@ export class PRReviewerService {
               const docProvider = await this.getDocProvider();
               const storiesList: string[] = [];
 
-              for (const ref of workItemRefs) {
-                if (ref.id) {
+              for (const refId of workItemRefs) {
+                if (refId) {
                   try {
                     let ticketData: TicketData;
                     if (issueTracker) {
-                      ticketData = await issueTracker.fetchTicket(ref.id);
+                      ticketData = await issueTracker.fetchTicket(refId);
                     } else {
                       ticketData = {
-                        id: ref.id,
-                        title: `Work Item ${ref.id}`,
+                        id: refId,
+                        title: `Work Item ${refId}`,
                         description: '',
                       };
                     }
@@ -1000,7 +733,7 @@ export class PRReviewerService {
                     }
                   } catch (err) {
                     console.error(
-                      `Failed to fetch details for work item ${ref.id}:`,
+                      `Failed to fetch details for work item ${refId}:`,
                       err,
                     );
                   }
@@ -1311,108 +1044,13 @@ export class PRReviewerService {
       line?: number;
       comment: string;
     },
-    settings: AppSettings,
+    _settings: AppSettings,
   ): Promise<void> {
-    let prNumber = parseInt(prUrlOrId);
-    const azureConn = settings.connectors?.azureDevOps;
-    let org = azureConn?.org || '';
-    let project = azureConn?.project || '';
-
-    const parsedUrl = this.parsePRUrl(prUrlOrId);
-    if (parsedUrl) {
-      prNumber = parsedUrl.prNumber;
-      org = parsedUrl.org;
-      project = parsedUrl.project;
-    } else if (isNaN(prNumber)) {
-      throw new Error(`Invalid Pull Request URL or ID format: "${prUrlOrId}"`);
-    } else {
-      // Try to detect org and project from remote URL
-      const remoteUrl = await this.gitService.getRemoteUrl(repoPath);
-      if (remoteUrl) {
-        const parsedRemote = this.parseRemoteUrl(remoteUrl);
-        if (parsedRemote) {
-          org = org || parsedRemote.org;
-          project = project || parsedRemote.project;
-        }
-      }
+    const provider = await this.getCodeReviewProvider();
+    if (!provider) {
+      throw new Error('No code review provider configured.');
     }
-
-    if (!org) {
-      throw new Error(
-        'Azure DevOps Organization is not configured in settings and could not be detected from git remote.',
-      );
-    }
-    if (!project) {
-      throw new Error(
-        'Azure DevOps Project is not configured in settings and could not be detected from git remote.',
-      );
-    }
-
-    const pat = settings.connectors?.azureDevOps?.pat;
-    if (!pat) {
-      throw new Error(
-        'Azure DevOps PAT token is missing. Please configure it in Settings.',
-      );
-    }
-
-    const orgUrl = this.getOrgUrl(org);
-    const authHandler = azdev.getPersonalAccessTokenHandler(pat);
-    const connection = new azdev.WebApi(orgUrl, authHandler);
-    const gitApi: IGitApi = await connection.getGitApi();
-
-    // Fetch the pull request to get the repository ID
-    const prDetails = await gitApi.getPullRequestById(prNumber);
-    if (!prDetails || !prDetails.repository || !prDetails.repository.id) {
-      throw new Error(`Pull Request #${prNumber} not found.`);
-    }
-
-    const repositoryId = prDetails.repository.id;
-
-    const disclaimer = [
-      '',
-      '> Generated with Stitch and GitHub Copilot.',
-      '> Like any AI generated content, mistakes and hallucinations can occur. Please review before relying on it.',
-    ].join('\n');
-
-    const contentWithDisclaimer = comment.comment + disclaimer;
-
-    // Define the thread
-    const thread: GitPullRequestCommentThread = {
-      comments: [
-        {
-          parentCommentId: 0,
-          content: contentWithDisclaimer,
-          commentType: 1, // Text comment
-        },
-      ],
-      status: 1, // Active
-    };
-
-    if (comment.type === 'line' && comment.file && comment.line) {
-      let formattedPath = comment.file.replace(/\\/g, '/');
-      if (!formattedPath.startsWith('/')) {
-        formattedPath = '/' + formattedPath;
-      }
-      thread.threadContext = {
-        filePath: formattedPath,
-        rightFileStart: {
-          line: comment.line,
-          offset: 1,
-        },
-        rightFileEnd: {
-          line: comment.line + 1,
-          offset: 1,
-        },
-      };
-    }
-
-    try {
-      await gitApi.createThread(thread, repositoryId, prNumber, project);
-    } catch (error: unknown) {
-      const errMsg = error instanceof Error ? error.message : String(error);
-      throw new Error(`Failed to create comment thread: ${errMsg}`, {
-        cause: error,
-      });
-    }
+    const remoteUrl = await this.gitService.getRemoteUrl(repoPath);
+    await provider.postPRComment(repoPath, prUrlOrId, comment, remoteUrl);
   }
 }
